LPS-57210 Rename web keys to match patterns
