+++ modules/apps/mobile-device-rules/mobile-device-rules-web/src/META-INF/resources/rule/simple_rule.jsp
@@ -17,7 +17,7 @@
 <%@ include file="/init.jsp" %>
 
 <%
-MDRRule rule = (MDRRule)request.getAttribute(MobileDeviceRulesWebKeys.MOBILE_DEVICE_RULES_RULE);
+MDRRule rule = (MDRRule)request.getAttribute(MDRWebKeys.MOBILE_DEVICE_RULES_RULE);
 
 Set<String> operatingSystems = Collections.emptySet();
 int tablet = 0;
