+++ modules/apps/mobile-device-rules/mobile-device-rules-web/src/com/liferay/mobile/device/rules/web/portlet/action/EditActionMVCResourceCommand.java
@@ -15,7 +15,7 @@
 package com.liferay.mobile.device.rules.web.portlet.action;
 
 import com.liferay.mobile.device.rules.constants.MobileDeviceRulesPortletKeys;
-import com.liferay.mobile.device.rules.web.util.MobileDeviceRulesWebKeys;
+import com.liferay.mobile.device.rules.web.constants.MDRWebKeys;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCResourceCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
 import com.liferay.portal.kernel.util.ParamUtil;
@@ -50,8 +50,7 @@ public class EditActionMVCResourceCommand extends BaseMVCResourceCommand {
 		MDRAction action = MDRActionServiceUtil.fetchAction(actionId);
 
 		resourceRequest.setAttribute(
-			MobileDeviceRulesWebKeys.MOBILE_DEVICE_RULES_RULE_GROUP_ACTION,
-			action);
+			MDRWebKeys.MOBILE_DEVICE_RULES_RULE_GROUP_ACTION, action);
 
 		String type = ParamUtil.getString(resourceRequest, "type");
